% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_r6_finance_data.R
\name{FinanceData}
\alias{FinanceData}
\title{FinanceData Class for Managing Financial Data Tables}
\description{
FinanceData Class for Managing Financial Data Tables

FinanceData Class for Managing Financial Data Tables
}
\details{
The `FinanceData` class is designed to manage individual financial data tables (e.g., accounts, postings, rules)
in a SQLite database. This class provides methods to append, update, and retrieve data from these tables.


The function converts any columns that end with "date" to character format to ensure proper insertion into the database.
It then appends the rows to the table in-place.



The function converts any columns that end with "date" to character format to ensure proper insertion into the database.
It then updates the rows in the table, matching rows by their `id` column.
Unmatched rows are ignored, and the operation will be performed in-place.



The function collects the data from the database and converts any columns that end with "date" from character
format back to `Date` objects, using the format `%Y-%m-%d`.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Reference to the table in the database.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FinanceData-new}{\code{FinanceData$new()}}
\item \href{#method-FinanceData-rows_append}{\code{FinanceData$rows_append()}}
\item \href{#method-FinanceData-rows_update}{\code{FinanceData$rows_update()}}
\item \href{#method-FinanceData-get_data}{\code{FinanceData$get_data()}}
\item \href{#method-FinanceData-clone}{\code{FinanceData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FinanceData-new"></a>}}
\if{latex}{\out{\hypertarget{method-FinanceData-new}{}}}
\subsection{Method \code{new()}}{
Initialize the `FinanceData` object by connecting to a specific table in the database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FinanceData$new(src, from)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{src}}{The source database connection (e.g., a DBI connection object).}

\item{\code{from}}{The name of the table in the database that this object will manage.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FinanceData-rows_append"></a>}}
\if{latex}{\out{\hypertarget{method-FinanceData-rows_append}{}}}
\subsection{Method \code{rows_append()}}{
Append new rows to the database table managed by the `FinanceData` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FinanceData$rows_append(y, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{A tibble containing the rows to append.}

\item{\code{...}}{Additional arguments passed to `dplyr::rows_append`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FinanceData-rows_update"></a>}}
\if{latex}{\out{\hypertarget{method-FinanceData-rows_update}{}}}
\subsection{Method \code{rows_update()}}{
Update existing rows in the database table managed by the `FinanceData` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FinanceData$rows_update(y, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{A tibble containing the rows to update.}

\item{\code{...}}{Additional arguments passed to `dplyr::rows_update`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FinanceData-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-FinanceData-get_data}{}}}
\subsection{Method \code{get_data()}}{
Retrieve all data from the table managed by the `FinanceData` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FinanceData$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A tibble containing all rows from the table with date columns as `Date` objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FinanceData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FinanceData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FinanceData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
