% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_r6_ledger.R
\name{Ledger}
\alias{Ledger}
\title{Ledger Class for Managing Finance Data}
\description{
Ledger Class for Managing Finance Data

Ledger Class for Managing Finance Data
}
\details{
The `Ledger` class is designed to manage financial transactions, accounts, and rules
using a SQLite database. This class provides methods to initialize the database,
handle account and transaction data, and apply rules for categorizing transactions.


The function performs the following steps:
- Joins the `postings` table with the `accounts` table to get the account names.
- Left-joins with the `rules` table to categorize transactions based on `payee_name` and `description`.
- Transactions that aren't matched to any rule are categorized as "Income: Unknown" or "Expenses: Unknown" based on the transaction amount.
- Dates are converted from character format to `Date` objects.



The function performs the following steps:
- Selects relevant columns from the postings data.
- Extracts account information and either adds a new account or updates an existing account in the ledger.
- Appends the postings data to the `postings` table in the ledger.
- Missing values in character columns are replaced with empty strings.



The function performs the following steps:
- Selects relevant columns from the postings data.
- Missing values in character columns are replaced with empty strings.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{con}}{Database connection object}

\item{\code{accounts}}{Holds account data}

\item{\code{postings}}{Holds transaction data (postings)}

\item{\code{rules}}{Holds rules for categorizing transactions}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Ledger-new}{\code{Ledger$new()}}
\item \href{#method-Ledger-get_transactions}{\code{Ledger$get_transactions()}}
\item \href{#method-Ledger-import_postings}{\code{Ledger$import_postings()}}
\item \href{#method-Ledger-import_rules}{\code{Ledger$import_rules()}}
\item \href{#method-Ledger-finalize}{\code{Ledger$finalize()}}
\item \href{#method-Ledger-clone}{\code{Ledger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ledger-new"></a>}}
\if{latex}{\out{\hypertarget{method-Ledger-new}{}}}
\subsection{Method \code{new()}}{
Initialize the `Ledger` object by connecting to a SQLite database
and creating the necessary tables if they don't exist.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ledger$new(db_path = "~/ledger.sql")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db_path}}{The path to the SQLite database file. Defaults to "ledger.sql".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ledger-get_transactions"></a>}}
\if{latex}{\out{\hypertarget{method-Ledger-get_transactions}{}}}
\subsection{Method \code{get_transactions()}}{
Retrieve transactions (postings) from the ledger, joined with
account information and categorized using the rules table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ledger$get_transactions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A tibble containing transactions with account information and categorized data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ledger-import_postings"></a>}}
\if{latex}{\out{\hypertarget{method-Ledger-import_postings}{}}}
\subsection{Method \code{import_postings()}}{
Import postings from a processed CSV file into the ledger system.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ledger$import_postings(postings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{postings}}{A tibble containing processed postings data from the CSV file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ledger-import_rules"></a>}}
\if{latex}{\out{\hypertarget{method-Ledger-import_rules}{}}}
\subsection{Method \code{import_rules()}}{
Import rules from a CSV file into the ledger system.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ledger$import_rules(rules)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rules}}{A tibble containing rules data from the CSV file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ledger-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-Ledger-finalize}{}}}
\subsection{Method \code{finalize()}}{
Disconnect the database connection when the `Ledger` object is finalized.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ledger$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ledger-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Ledger-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ledger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
